Add in loads server for simplified firmware upgrade
The template.xml file was modified to add in a loads server so the phones can get firmware upgrades over the air and not require the provisioning server to have the firmware files, which take up a lot of space. Adding in the CSP Loads server allows the phones to get the latest firmware possible.